Move attributes utils to attributes directory
