+++ src/attributes/utils/data.ts
@@ -1,0 +1,67 @@
+import { FileUpload } from "@saleor/files/types/FileUpload";
+import { UploadErrorFragment } from "@saleor/fragments/types/UploadErrorFragment";
+import { FormsetData } from "@saleor/hooks/useFormset";
+import { AttributeValueInput } from "@saleor/types/globalTypes";
+import { MutationFetchResult } from "react-apollo";
+
+export const mergeFileUploadErrors = (
+  uploadFilesResult: Array<MutationFetchResult<FileUpload>>
+): UploadErrorFragment[] =>
+  uploadFilesResult.reduce((errors, uploadFileResult) => {
+    const uploadErrors = uploadFileResult.data.fileUpload.uploadErrors;
+    if (uploadErrors) {
+      return [...errors, ...uploadErrors];
+    }
+    return errors;
+  }, []);
+
+export const getFileValuesToUploadFromAttributes = (
+  attributesWithNewFileValue: FormsetData<null, File>
+) => attributesWithNewFileValue.filter(fileAttribute => !!fileAttribute.value);
+
+export const getFileValuesRemovedFromAttributes = (
+  attributesWithNewFileValue: FormsetData<null, File>
+) => attributesWithNewFileValue.filter(attribute => !attribute.value);
+
+export const getAttributesOfRemovedFiles = (
+  fileAttributesRemoved: FormsetData<null, File>
+) =>
+  fileAttributesRemoved.map(attribute => ({
+    file: undefined,
+    id: attribute.id,
+    values: []
+  }));
+
+export const getAttributesOfUploadedFiles = (
+  fileValuesToUpload: FormsetData<null, File>,
+  uploadFilesResult: Array<MutationFetchResult<FileUpload>>
+) =>
+  uploadFilesResult.map((uploadFileResult, index) => {
+    const attribute = fileValuesToUpload[index];
+
+    return {
+      file: uploadFileResult.data.fileUpload.uploadedFile.url,
+      id: attribute.id,
+      values: []
+    };
+  });
+
+export const getAttributesFromFileUploadResult = (
+  attributesWithNewFileValue: FormsetData<null, File>,
+  uploadFilesResult: Array<MutationFetchResult<FileUpload>>
+): AttributeValueInput[] => {
+  const removedFileValues = getFileValuesRemovedFromAttributes(
+    attributesWithNewFileValue
+  );
+  const fileValuesToUpload = getFileValuesToUploadFromAttributes(
+    attributesWithNewFileValue
+  );
+
+  const removedFileAttributes = getAttributesOfRemovedFiles(removedFileValues);
+  const uploadedFileAttributes = getAttributesOfUploadedFiles(
+    fileValuesToUpload,
+    uploadFilesResult
+  );
+
+  return uploadedFileAttributes.concat(removedFileAttributes);
+};
