+++ src/pages/utils/handlers.ts
@@ -1,25 +1,10 @@
-import {
-  AttributeInput,
-  AttributeInputData
-} from "@saleor/components/Attributes";
-import {
-  FileUpload,
-  FileUploadVariables
-} from "@saleor/files/types/FileUpload";
+import { AttributeInputData } from "@saleor/components/Attributes";
 import { FormChange } from "@saleor/hooks/useForm";
 import { FormsetChange, FormsetData } from "@saleor/hooks/useFormset";
-import {
-  AttributeInputTypeEnum,
-  AttributeValueInput
-} from "@saleor/types/globalTypes";
 import { toggle } from "@saleor/utils/lists";
-import { MutationFetchResult } from "react-apollo";
 
 import { PageDetails_page_pageType } from "../types/PageDetails";
-import {
-  getAttributeInputFromPageType,
-  getFileValuesToUploadFromAttributes
-} from "./data";
+import { getAttributeInputFromPageType } from "./data";
 
 export function createPageTypeSelectHandler(
   change: FormChange,
@@ -69,51 +54,3 @@ export function createAttributeMultiChangeHandler(
     changeAttributeData(attributeId, newAttributeValues);
   };
 }
-
-interface PageAttributesArgs {
-  attributes: AttributeInput[];
-  attributesWithAddedNewFiles: AttributeValueInput[];
-}
-
-export const prepareAttributesInput = ({
-  attributes,
-  attributesWithAddedNewFiles
-}: PageAttributesArgs): AttributeValueInput[] =>
-  attributes.map(attribute => {
-    if (attribute.data.inputType === AttributeInputTypeEnum.FILE) {
-      const attributeWithNewFile = attributesWithAddedNewFiles.find(
-        attributeWithNewFile => attribute.id === attributeWithNewFile.id
-      );
-      if (attributeWithNewFile) {
-        return {
-          file: attributeWithNewFile.file,
-          id: attributeWithNewFile.id
-        };
-      }
-      return {
-        file:
-          attribute.data.selectedValues &&
-          attribute.data.selectedValues[0]?.file?.url,
-        id: attribute.id
-      };
-    }
-    return {
-      id: attribute.id,
-      values: attribute.value[0] === "" ? [] : attribute.value
-    };
-  });
-
-export const handleUploadMultipleFiles = async (
-  attributesWithNewFileValue: FormsetData<null, File>,
-  uploadFile: (
-    variables: FileUploadVariables
-  ) => Promise<MutationFetchResult<FileUpload>>
-) =>
-  Promise.all(
-    getFileValuesToUploadFromAttributes(attributesWithNewFileValue).map(
-      fileAttribute =>
-        uploadFile({
-          file: fileAttribute.value
-        })
-    )
-  );
